% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{aggs}
\alias{aggs}
\title{Define Elasticsearch aggregation.}
\usage{
aggs(json)
}
\arguments{
\item{json}{JSON object describing the aggregation that needs to be executed.}
}
\value{
An \code{elastic_aggs} object.
}
\description{
Define Elasticsearch aggregation.
}
\examples{
avg_sepal_width_per_cat <- aggs('{"avg_sepal_width_per_cat": {
      "terms": {"field": "species"},
      "aggs": {"avg_sepal_width": {"avg": {"field": "sepal_width"}}}}
}')
}
\seealso{
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations.html}
}

