% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\name{elastic_aggregation}
\alias{elastic_aggregation}
\title{Execute a simple aggregation and return the results in a data frame.}
\usage{
elastic_aggregation(elasticsearch, agg_def, agg_metric, query_def = list(query
  = list(match_all = list())))
}
\arguments{
\item{elasticsearch}{connection details for the Elasticsearch server we want to access (and optionally the index and type too).}

\item{agg_def}{aggregation definition using nested list structure.}

\item{agg_metric}{the metrics we are interested in computing.}

\item{query_def}{query/filter definition using nested list structure.}
}
\value{
data frame of aggregation results.
}
\description{
This is experimental and only supports a single pre-aggregation query.
}
\examples{
\dontrun{
time_buckets <- list('agg_time' = list('date_histogram' = list('field' = 'accept_date', 'interval' = 'hour' )))
avg_and_max_responses <- list( 'avg_response' = list('avg' = list('field' = 'Tt')), 'max_response' = list('max' = list('field' = 'Tt')) )

elastic_aggregation(elastic_conn(...), time_buckets, avg_and_max_responses)
}
}

