% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\name{elastic_bulk_reingest}
\alias{elastic_bulk_reingest}
\title{Re-ingest an index into Elasticsearch with a new mapping.}
\usage{
elastic_bulk_reingest(elasticsearch, mapping = "")
}
\arguments{
\item{elasticsearch}{connection details for the Elasticsearch server we want to access (and optionally the index and type too).}

\item{mapping}{the new mapping to upload prior to re-ingesting the index (must be in raw JSON format as specified by Elasticsearch).}
}
\value{
TRUE or FALSE depending on the success of the operations.
}
\description{
Index mappings for existing fields cannot be changed. For example, it is not possible to to change a field from 'analyzed' to 'not_analyzed' once the mapping
for the field has been defined. \code{elastic_bulk_reingest} automates the laborious process of downloading the contents of an index, uploading a revised mapping,
and then re-loading the index.
}

