% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{elastic_bulk_size}
\alias{elastic_bulk_size}
\title{Get or Set the elasticsearch bulk size.}
\usage{
elastic_bulk_size(size = NULL)
}
\arguments{
\item{size}{integer The number of document to bulk per request.}
}
\description{
Get or Set the elasticsearch bulk size.
}
\examples{
elastic_bulk_size(100)
}
\seealso{
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-bulk.html}
}
