% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\name{elastic_create_dated_index_name}
\alias{elastic_create_dated_index_name}
\title{Create a valid Elasticsearch date-based index name.}
\usage{
elastic_create_dated_index_name(elastic_index_name, date)
}
\arguments{
\item{elastic_index_name}{the base (or stem) part of the index name - e.g. the 'logs' part of 'logs-2015.01.01'.}

\item{date}{POSIXct object or a string describing the date with the format YYYY-MM-DD.}
}
\value{
A charcter string representing a valid Elasticsearch date-based index name.
}
\description{
Create a valid Elasticsearch date-based index name.
}
\examples{
elastic_create_dated_index_name('logs', '2015-01-31')
elastic_create_dated_index_name('logs', Sys.time())
}

