% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\name{elastic_create_new_index}
\alias{elastic_create_new_index}
\title{Create a new index.}
\usage{
elastic_create_new_index(elasticsearch, mapping = "")
}
\arguments{
\item{elasticsearch}{connection details for the Elasticsearch server we want to access (and optionally the index and type too).}

\item{mapping}{optional mapping to upload (must be in raw JSON format as specified by Elasticsearch).}
}
\value{
TRUE or FALSE depending on the success of the operations.
}
\description{
Create a new index.
}

