% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{elastic_parallel_index}
\alias{elastic_parallel_index}
\title{Get or Set the elasticsearch parallel index.}
\usage{
elastic_parallel_index(parallel = NULL)
}
\arguments{
\item{parallel}{boolean The indicator of parallel use.}
}
\description{
Get or Set the elasticsearch parallel index.
}
\examples{
elastic_parallel_index(TRUE)
}
