% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{elastic_predicates}
\alias{elastic_predicates}
\alias{is_elastic}
\alias{is_elastic_aggs}
\alias{is_elastic_api}
\alias{is_elastic_query}
\alias{is_elastic_rescource}
\alias{is_elastic_sort}
\title{elasticsearchr predicate functions.}
\usage{
is_elastic(x)

is_elastic_rescource(x)

is_elastic_api(x)

is_elastic_query(x)

is_elastic_aggs(x)

is_elastic_sort(x)
}
\arguments{
\item{x}{An elasticsearchr object.}
}
\value{
Boolean.
}
\description{
Predicate functions for identifying different elasticsearchr object types.
}

