% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\name{elastic_query}
\alias{elastic_query}
\title{Execute a simple query (or filter) and return the results in a data frame or CSV file.}
\usage{
elastic_query(elasticsearch, query_def, to_file = FALSE,
  file_name = paste0("elastic_query--", Sys.time(), ".csv"),
  to_file_and_df = FALSE)
}
\arguments{
\item{elasticsearch}{connection details for the Elasticsearch server we want to access (and optionally the index and type too).}

\item{query_def}{query/filter definition using nested list structure.}

\item{to_file}{write the result to a CSV file instead of a data frame?}

\item{file_name}{if \code{to_file = TRUE} then a file name can be supplied.}

\item{to_file_and_df}{return a data frame as well as a CSV file?}
}
\value{
data frame of query results or NULL.
}
\description{
This is experimental. Currently, there is support for combining one query with one filter. Multiple filters and queries will come in a later version.
}
\examples{
\dontrun{
elastic_query(elastic_conn_(...), query_def = list('filter' = list('range' = list("Tt" = list("gt"=2000)))) )

query = list('query' = list('term' = list("status_code"="404")), 'filter' = list('range' = list("feconn" = list("gt"=30))))
elastic_query(elastic_conn_(...), query)
}
}

