% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elastic_version}
\alias{elastic_version}
\title{Elasticsearch version}
\usage{
elastic_version(url)
}
\arguments{
\item{url}{A valid URL to an Elasticsearch cluster.}
}
\value{
A list with the \code{major}, \code{minor} and \code{build} numbers.
}
\description{
Returns the major, minor and build version numbers for an Elasticsearch cluster, given a valid
URL to an Elasticsearch cluster.
}
\examples{
\dontrun{
elastic_version("http://localhost:9200")
$major
[1] 5

$minor
[1] 0

$build
[1] 1
}
}

