% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\docType{package}
\name{elasticr}
\alias{elasticr}
\alias{elasticr-package}
\title{elasticr: A package for making working with Elasticsearch easy from R.}
\description{
elasticr: A package for making working with Elasticsearch easy from R.
}
\section{elasticr functions}{

elastic_conn
}

