% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\docType{package}
\name{elasticsearchr}
\alias{elasticsearchr}
\alias{elasticsearchr-package}
\title{elasticsearchr: A package for making working with Elasticsearch easy from R.}
\description{
elasticsearchr: A package for making working with Elasticsearch easy from R.
}
\section{elasticsearchr functions}{

elastic_conn
}

