% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_query_results}
\alias{extract_query_results}
\alias{extract_query_results}
\alias{extract_aggs_results}
\alias{extract_id_results}
\title{Elasticsearch HTTP response data extraction functions.}
\usage{
extract_query_results(response)

extract_aggs_results(response)

extract_id_results(response)
}
\arguments{
\item{response}{An HTTP response from a response to a search API request.}
}
\value{
A data.frame of response results.
}
\description{
Functions for extracting the different types of data that can be contained in a response to a
search API request.
}
