% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{\%delete\%}
\alias{\%delete\%}
\title{Delete Elasticsearch index.}
\usage{
rescource \%delete\% approve
}
\arguments{
\item{rescource}{An \code{elastic_rescource} object that contains the information on the
Elasticsearch cluster, index and document type, where the indexed data will reside. If this does
not already exist, it will be created automatically.}

\item{approve}{Must be equal to \code{"TRUE"} for deletion for all documents in a rescource,
OR be a character vector of document ids if only specific documents need to be deleted.}
}
\description{
Delete all of the documents within a particular document type (if specified), or delete an
entire index (if the document type is unspecified.)
}
\examples{
\dontrun{
elastic("http://localhost:9200", "iris", "data") \%delete\% TRUE
}
}
