% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{\%info\%}
\alias{\%info\%}
\title{Get cluster and index (meta) data.}
\usage{
rescource \%info\% info
}
\arguments{
\item{rescource}{An \code{elastic_rescource} object that contains the information on the
Elasticsearch cluster, index and document type, where the indexed data will reside. If this does
not already exist, it will be created automatically.}

\item{info}{\code{elastic_info} object.}
}
\description{
An operator to be used with requests for information
}
\examples{
\dontrun{
elastic("http://localhost:9200", "iris", "data") \%info\% list_indices()
elastic("http://localhost:9200", "iris", "data") \%info\% list_fields()
}
}
