% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{\%search\%}
\alias{\%search\%}
\title{Execute query or search.}
\usage{
rescource \%search\% search
}
\arguments{
\item{rescource}{An \code{elastic_rescource} object that contains the information on the
Elasticsearch cluster, index and document type, where the indexed data will reside. If this does
not already exist, it will be created automatically.}

\item{search}{\code{elastic_query} or \code{elastic_aggs} object.}
}
\value{
A data.frame of search or aggregation results.
}
\description{
Execute query or search.
}
\examples{
\dontrun{
results <- elastic("http://localhost:9200", "iris", "data") \%search\% query('{"match_all": {}}')
head(results)
#   sepal_length sepal_width petal_length petal_width species
# 1          4.8         3.0          1.4         0.1  setosa
# 2          4.3         3.0          1.1         0.1  setosa
# 3          5.8         4.0          1.2         0.2  setosa
# 4          5.1         3.5          1.4         0.3  setosa
# 5          5.2         3.5          1.5         0.2  setosa
# 6          5.2         3.4          1.4         0.2  setosa
}
}

