% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_fields}
\alias{list_fields}
\title{List of fields in index information.}
\usage{
list_fields()
}
\value{
An \code{elastic_info} object.
}
\description{
List of fields in index information.
}
\examples{
list_fields()
}
