% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_indices}
\alias{list_indices}
\title{List of indices in cluster information.}
\usage{
list_indices()
}
\value{
An \code{elastic_info} object.
}
\description{
List of indices in cluster information.
}
\examples{
list_indices()
}
