% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{mapping_default_simple}
\alias{mapping_default_simple}
\title{Simple Elasticsearch default mappings for non-text-search analytics}
\usage{
mapping_default_simple()
}
\description{
This mapping switches-off the text analyser for all fields of type 'string' (i.e. switches off
free text search), allows all text search to work with case-insensitive lowercase terms, and
maps any field with the name 'timestamp' to type 'date', so long as it has the appropriate
string or long format.
}

