% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{mapping_fielddata_true}
\alias{mapping_fielddata_true}
\title{Elasticsearch 5.x default mappings enabling fielddata for text fields}
\usage{
mapping_fielddata_true()
}
\description{
A default mapping that enables fielddata for all string/text fields in Elasticsearch 5.x.
}
