% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query}
\alias{query}
\title{Define Elasticsearch query.}
\usage{
query(json, size = 0)
}
\arguments{
\item{json}{JSON object describing the query that needs to be executed.}

\item{size}{[optional] The number of documents to return. If left unspecified, then the default
if to return all documents.}
}
\value{
An \code{elastic_query} object.
}
\description{
Define Elasticsearch query.
}
\examples{
all_docs <- query('{"match_all": {}}')
}
\seealso{
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl.html}
}
