% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scroll_search}
\alias{scroll_search}
\title{Title}
\usage{
scroll_search(rescource, api_call_payload,
  extract_function = extract_query_results)
}
\arguments{
\item{extract_function}{}
}
\description{
Title
}

