% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{select_fields}
\alias{select_fields}
\title{Define Elasticsearch query source filter.}
\usage{
select_fields(json)
}
\arguments{
\item{json}{JSON object describing the aggregation that needs to be executed.}
}
\value{
An \code{elastic_source_filter} object.
}
\description{
Define Elasticsearch query source filter.
}
\seealso{
\url{https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-source-filtering.html}
}
