% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sort}
\alias{sort}
\title{Define Elasticsearch query sort}
\usage{
sort(json)
}
\arguments{
\item{json}{JSON object describing the sorting required on the query results.}
}
\value{
An \code{elastic_sort} object.
}
\description{
Define Elasticsearch query sort
}
\examples{
sort_by_key <- sort('[{"sort_key": {"order": "asc"}}]')
}
\seealso{
\url{https://www.elastic.co/guide/en/elasticsearch/reference/5.0/search-request-sort.html}
}

