% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_json}
\alias{valid_json}
\title{Valid JSON string predicate function}
\usage{
valid_json(json)
}
\arguments{
\item{json}{Candidate JSON object as a string.}
}
\value{
Boolean.
}
\description{
Valid JSON string predicate function
}
\examples{
\dontrun{
good_json <- '{"id": 1}'
valid_json(good_json)
# TRUE

bad_json <- '{"id": 1a}'
valid_json(bad_json)
# FALSE
}
}

