% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elastic_functions.R
\name{elastic_conn}
\alias{elastic_conn}
\title{Define a 'connection' to an Elasticsearch server}
\usage{
elastic_conn(url = "http://localhost:9200", index = "", type = "")
}
\arguments{
\item{url}{Location of the Elasticsearch serves that contains the index and type we're interestd in working with.}

\item{index}{Name of the index we want to work with.}

\item{type}{Name of the type we want to work with (optional).}
}
\value{
A list containing the the location of the Elasticsearch server, the index and type (option) that we want to work with (as well as the index/type path).
}
\description{
Define a 'connection' to an Elasticsearch server
}
\examples{
elastic_conn('http://localhost:9200', 'analytics', 'users')
}

