% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchr.R
\docType{package}
\name{elasticsearchr}
\alias{elasticsearchr}
\alias{elasticsearchr-package}
\title{elasticsearchr: a lightweight Elasticsearch client for R.}
\description{
Allows you to index, update and delete documents as well as run queries and aggregations.
}

